requirements.txt
# Python dependencies for the big data platform
# These are installed automatically in Docker containers

# Airflow providers
apache-airflow-providers-apache-spark>=4.0.0
apache-airflow-providers-mysql>=5.0.0
apache-airflow-providers-apache-hive>=6.0.0

# Database connectors
mysql-connector-python>=8.0.33
psycopg2-binary>=2.9.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# HTTP requests
requests>=2.28.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0